+++ Engine/Source/Renderer/Buffer/Buffer.cpp
@@ -5,6 +5,8 @@ namespace Core
 {
     Buffer::Buffer(Type _type)
     {
+        id = 0;
+        stride = 0;
         type = _type;
 
         glGenBuffers(1, &id);
@@ -19,7 +21,26 @@ namespace Core
 
     void Buffer::Bind()
     {
-        glBindBuffer(type == Vertex ? GL_ARRAY_BUFFER : GL_ELEMENT_ARRAY_BUFFER, id);
+
+        switch (type)
+        {
+        case Vertex:
+
+            glBindBuffer(GL_ARRAY_BUFFER, id);
+            for (CeU32 i = 0; i < layouts.size(); i++)
+            {
+                CeU32 l = layouts[i].location;
+                CeU32 o = layouts[i].offset;
+                CeU32 s = layouts[i].size;
+                glEnableVertexAttribArray(l);
+                glVertexAttribPointer(l, s, GL_FLOAT, GL_FALSE, stride * sizeof(float), (void *)(o * sizeof(float)));
+            }
+            break;
+
+        case Index:
+            glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, id);
+            break;
+        }
     }
 
     void Buffer::Unbind()
@@ -32,7 +53,7 @@ namespace Core
         switch (type)
         {
         case Vertex:
-            glDrawArrays(GL_TRIANGLES, 0, 3);
+            glDrawArrays(GL_TRIANGLES, 0, size / stride);
             break;
 
         case Index:
@@ -45,5 +66,14 @@ namespace Core
     {
         size = _size;
         glBufferData(type == Vertex ? GL_ARRAY_BUFFER : GL_ELEMENT_ARRAY_BUFFER, _size, data, GL_STATIC_DRAW);
+    }
+
+    void Buffer::AddLayout(int location, int offset, int size)
+    {
+        if (type != Vertex)
+            return;
+
+        stride += size;
+        layouts.push_back({location, offset, size});
     }
 }