+++ Engine/Source/Renderer/Buffer/Buffer.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "Base.h"
+#include <vector>
 
 namespace Core
 {
@@ -14,9 +15,21 @@ namespace Core
         };
 
     private:
+        struct Layout
+        {
+            int location;
+            int offset;
+            int size;
+
+            Layout(int l, int o, int s) : location(l), offset(o), size(s) {}
+        };
+
         Type type;
         CeU32 id;
         int size;
+
+        std::vector<Layout> layouts;
+        int stride;
 
     public:
         Buffer(Type type);
@@ -26,7 +39,9 @@ namespace Core
         void Unbind();
         void Draw();
 
-        void Upload(void* data, int size);
+        void Upload(void *data, int size);
+
+        void AddLayout(int location, int offset, int size);
 
         inline Type GetType() { return type; };
         inline CeU32 GetId() { return id; };